Virtual Expression
This codes uses the DigiPot on the Arduino shield as a virtual expression pedal. Connect a TRS jack between the board and your desired (guitar) effect pedal. Footswitch 0 activates the static mode: expression in (analog pin 5) = expression out on the DigiPot. Footswitch 1 starts a sawtooth wave shape, 2 a triangular wave shape and 3 a square wave shape. The expression input sets the rate for the wave shapes. When the Arduine receives MIDI clock data, the wave shapes are synced to the beat, with a cycle length of either 1, 2 or 4 beats. On MIDI tempo, wave shapes can be reversed when you press and hold footswitch 0 and press one of the others. Without MIDI, reverse is only available on the sawtooth wave shape. Static mode is always independent from MIDI tempo.