Freeze Footcontroller
A program to control two Electro Harmonix Freezes via an external footswitch, controlled by the relays connected to the Arduino, and a Line6 HX Stomp or similar effect pedal with multiple effect loops controllable via Midi. It has 5 modes:
1: All Freeze Modes. Set the Freeze to fast mode, the Slow and Latch mode are mimicked by the footcontroller.
2: Solo and Mute. Holds two independent sound sources, mutes one of them to allow a solo from the other.
3: Dual Layer. Hold two layers of sounds and change them at will.
4: Triple Layer. Same as the Dual Layer but with a third one provided by the Helix's Poly Sustain effect.
5: Infinite Freeze. Freezes one sound, stores it along with the new input sound to the other and back again, creating infinite layers.

To change modes, either send Midi PC messages 1 to 5 (or 0 to 4) on Midi channel 16 or reset the power holding a footswitch. When none are pressed, mode 1 is activated.